pyproject.toml
[tool.poetry]
name = "python-app-template"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.4"

[tool.poetry.group.iac.dependencies]
aws-cdk-lib = "^2.135.0"
constructs = ">=10.0.0,<11.0.0"
aws-cdk-aws-lambda-python-alpha = "^2.135.0a0"

[tool.poetry.group.iac_dev.dependencies]
pytest = "^8.1.1"

[tool.poetry.group.server.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
